Split genserver and public API - client
